Refactor code to move authentication functionlity to CommonBusinessService.java file.
