+++ quora-service/src/main/java/com/upgrad/quora/service/business/CommonBusinessService.java
@@ -31,18 +31,40 @@ public class CommonBusinessService {
      */
     public UserEntity userProfile(String accessToken, String userId) throws AuthorizationFailedException, UserNotFoundException {
 
-        UserAuthTokenEntity userAuthTokenEntity = userAuthDao.getUserAuthByToken((accessToken));
+        final String signoutExceptionMessage = "User is signed out.Sign in first to get user details";
+        UserEntity authenticatedUser = getAuthenticatedUser(accessToken, signoutExceptionMessage);
+
+        //fetch user profile based on given user id.
+        UserEntity fetchedUserProfile = userDao.getUserByUuid(userId);
+
+        if (fetchedUserProfile == null) {
+            throw new UserNotFoundException("USR-001", "User with entered uuid does not exist");
+        }
+        return fetchedUserProfile;
+    }
+
+    /**
+     * This method fetches the authenticated user entity based on given authToken.
+     *
+     * @param authToken               The JWT access token of the user
+     * @param signoutExceptionMessage The Exception message to show for Exception ATHR-002
+     * @return UserEntity The user data of the authorized user
+     * @throws AuthorizationFailedException This exception is thrown, if the user is not signed in or it has signed out
+     */
+    public UserEntity getAuthenticatedUser(final String authToken, final String signoutExceptionMessage) throws AuthorizationFailedException {
+        UserAuthTokenEntity userAuthTokenEntity = userAuthDao.getUserAuthByToken(authToken);
+        //Check if the userAuthToken is not present in the database
         if (userAuthTokenEntity == null) {
             throw new AuthorizationFailedException("ATHR-001", "User has not signed in");
         }
+
+        //User is signed out if either JWT token is expired or user has logged out
         final ZonedDateTime now = ZonedDateTime.now();
-        if (userAuthTokenEntity.getLogoutAt() != null && userAuthTokenEntity.getLogoutAt().isBefore(now)) {
-            throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to get user details");
+        if (userAuthTokenEntity.getExpiresAt().isBefore(now) || userAuthTokenEntity.getLogoutAt() != null) {
+            throw new AuthorizationFailedException("ATHR-002", signoutExceptionMessage);
         }
-        UserEntity userEntity = userDao.getUserByUuid(userId);
-        if (userEntity == null) {
-            throw new UserNotFoundException("USR-001", "User with entered uuid does not exist");
-        }
-        return userEntity;
+
+        return userAuthTokenEntity.getUser();
     }
+
 }
